requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
transformers==4.35.2
torch==2.1.1
torchaudio==2.1.1
moviepy==1.0.3
ffmpeg-python==0.2.0
boto3==1.34.0
pytest==7.4.3
pytest-asyncio==0.21.1 